refactor(ui): split useTypingTest.ts into focused modules (989→596 lines)
Extract the pure, React-free layers into their own modules — word-supply
(config→word/quote resolution and the time-mode refill policy), run-state
(state shape + keystroke/word-advance reducers), romaji-input (romaji-mode
key semantics), and matrix-layers (matrix key→layer/keycode resolution) —
leaving the hook with its React wiring only. Code moved verbatim; the
public surface is unchanged via type re-exports, so no consumer or test
needed edits.
