+++ src/renderer/typing-test/run-state.ts
@@ -1,0 +1,175 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+/** Run state shape for a typing-test attempt, plus the pure reducers that
+ *  advance it on keystrokes, backspace, and word submits. */
+
+import type { TypingTestConfig, Quote } from './types'
+import { type WordsForConfig, createWordsForConfigSync, refillTimeModeWords } from './word-supply'
+
+export type TypingTestStatus = 'countdown' | 'waiting' | 'running' | 'finished' | 'paused'
+
+/** Check if a key is a word-submit key (half-width space or full-width space). */
+export function isSubmitKey(key: string): boolean {
+  return key === ' ' || key === '\u3000'
+}
+
+export interface WordResult {
+  word: string
+  typed: string
+  correct: boolean
+}
+
+export interface TypingTestState {
+  status: TypingTestStatus
+  /** Unique id for this run, regenerated by `freshState` on every
+   *  (re)start and preserved across pause/resume. Tags analytics events
+   *  so Analyze can slice a test material down to individual History
+   *  runs. */
+  runId: string
+  words: string[]
+  currentWordIndex: number
+  currentInput: string
+  compositionText: string
+  wordResults: WordResult[]
+  startTime: number | null
+  endTime: number | null
+  correctChars: number
+  incorrectChars: number
+  currentQuote: Quote | null
+  wpmHistory: number[]
+  /** Word indices that end a line (imported fileImport text only). At these
+   *  words Enter advances; elsewhere Space advances. Empty for every other
+   *  mode, so their submit behaviour is unchanged. */
+  lineBreaks: Set<number>
+  /** Leading whitespace per logical line (imported fileImport text, display only).
+   *  Indexed by line order; empty for every other mode. */
+  lineIndents: string[]
+  /** Romaji keystrokes accepted so far for the current word (romajiInput
+   *  mode only). Rejected keystrokes are never appended. Replayed through
+   *  a fresh `createRomajiMatcher` on every read/write instead of holding
+   *  a live matcher instance, so state transitions stay pure — see
+   *  `buildRomajiMatcher`. Reset to '' whenever the word advances. */
+  romajiKeystrokes: string
+}
+
+export function createInitialState(config: TypingTestConfig, language: string, status: TypingTestStatus = 'waiting'): TypingTestState {
+  return freshState(createWordsForConfigSync(config, language), status)
+}
+
+export function freshState({ words, quote, lineBreaks, lineIndents }: WordsForConfig, status: TypingTestStatus = 'waiting'): TypingTestState {
+  return {
+    status,
+    runId: crypto.randomUUID(),
+    words,
+    currentWordIndex: 0,
+    currentInput: '',
+    compositionText: '',
+    wordResults: [],
+    startTime: null,
+    endTime: null,
+    correctChars: 0,
+    incorrectChars: 0,
+    currentQuote: quote,
+    wpmHistory: [],
+    lineBreaks: new Set(lineBreaks),
+    lineIndents,
+    romajiKeystrokes: '',
+  }
+}
+
+export function handleChar(state: TypingTestState, char: string): TypingTestState {
+  if (state.currentWordIndex >= state.words.length) return state
+  return {
+    ...state,
+    currentInput: state.currentInput + char,
+  }
+}
+
+/** If the last word is fully typed, finalize it and finish the test. */
+export function tryFinishLastWord(state: TypingTestState): TypingTestState | null {
+  if (state.currentWordIndex !== state.words.length - 1) return null
+  const currentWord = state.words[state.currentWordIndex]
+  if (state.currentInput !== currentWord) return null
+
+  // Count chars without trailing space bonus (no space needed for last word)
+  let correct = 0
+  for (let i = 0; i < currentWord.length; i++) correct++
+
+  return {
+    ...state,
+    currentWordIndex: state.currentWordIndex + 1,
+    currentInput: '',
+    wordResults: [...state.wordResults, { word: currentWord, typed: currentWord, correct: true }],
+    correctChars: state.correctChars + correct,
+    incorrectChars: state.incorrectChars,
+    status: 'finished',
+    endTime: Date.now(),
+  }
+}
+
+/** Shared word-advance tail, run after a word has just been finalized into
+ *  `base.wordResults` and `base.currentWordIndex` moved past it. Used by
+ *  both the Space/Enter submit path (`handleSpace`) and the romaji
+ *  auto-advance path (`handleRomajiChar`) so the time-mode word-supply
+ *  refill and the words/quote finish transition stay in one place. */
+export function advanceAfterWord(base: TypingTestState, config: TypingTestConfig, language: string): TypingTestState {
+  const nextIndex = base.currentWordIndex
+
+  // Time mode: extend words if running low, never finish from words
+  if (config.mode === 'time') {
+    const refilled = refillTimeModeWords(base.words, nextIndex, config, language)
+    return refilled ? { ...base, words: refilled } : base
+  }
+
+  // Words and quote modes: finish when all words typed
+  if (nextIndex >= base.words.length) {
+    return { ...base, status: 'finished', endTime: Date.now() }
+  }
+  return base
+}
+
+export function handleSpace(state: TypingTestState, config: TypingTestConfig, language: string): TypingTestState {
+  if (state.currentWordIndex >= state.words.length) return state
+
+  const currentWord = state.words[state.currentWordIndex]
+  const typed = state.currentInput
+  const isCorrect = typed === currentWord
+  const charCounts = computeWordCharCounts(currentWord, typed)
+
+  const nextIndex = state.currentWordIndex + 1
+
+  const base: TypingTestState = {
+    ...state,
+    currentWordIndex: nextIndex,
+    currentInput: '',
+    wordResults: [...state.wordResults, { word: currentWord, typed, correct: isCorrect }],
+    correctChars: state.correctChars + charCounts.correct,
+    incorrectChars: state.incorrectChars + charCounts.incorrect,
+  }
+
+  return advanceAfterWord(base, config, language)
+}
+
+export function handleBackspace(state: TypingTestState): TypingTestState {
+  if (state.currentInput.length === 0) return state
+  return {
+    ...state,
+    currentInput: state.currentInput.slice(0, -1),
+  }
+}
+
+function computeWordCharCounts(word: string, typed: string): { correct: number; incorrect: number } {
+  const len = Math.max(typed.length, word.length)
+  let correct = 1 // count the space separator as a correct char
+  let incorrect = 0
+
+  for (let i = 0; i < len; i++) {
+    if (i < typed.length && i < word.length && typed[i] === word[i]) {
+      correct++
+    } else {
+      incorrect++
+    }
+  }
+
+  return { correct, incorrect }
+}
